new dataset

+++ grpc-server/main.py
@@ -8,6 +8,7 @@ import grpc
 import csv
 import logging
 import xml.etree.ElementTree as ET
+import xml.sax.saxutils as saxutils
 from xml.dom import minidom
 from lxml import etree
 import xmlschema
@@ -18,35 +19,61 @@ LOG_FORMAT = "%(asctime)s - %(name)s - %(levelname)s - %(message)s"
 logging.basicConfig(level=logging.INFO, format=LOG_FORMAT)
 logger = logging.getLogger("FileService")
 
+# =============== Functions =================
 
-# Helper Functions for CSV to XML conversion
+def csv_to_xml(csv_file, xml_file, objname, chunk_size=10000):
+    
+    if not os.path.exists(csv_file):
+        raise FileNotFoundError(f"Arquivo {csv_file} não encontrado.")
 
-def csv_to_xml(csv_file, xml_file, objname):
-    root = ET.Element("root")
-
-    with open(csv_file, mode='r', newline='', encoding='utf-8') as csvfile:
-        reader = csv.DictReader(csvfile, delimiter=',')
-        for line in reader:
-            obj = ET.Element(objname)
-            for field, value in line.items():
-                element = ET.SubElement(obj, field)
-                element.text = value
-            root.append(obj)
-
-    xml_string = ET.tostring(root, encoding='utf-8')
-    xml_pretty = minidom.parseString(xml_string).toprettyxml(indent='  ')
-
+    
     with open(xml_file, mode='w', encoding='utf-8') as xmlfile:
-        xmlfile.write(xml_pretty)
+        # initial root tag
+        xmlfile.write('<?xml version="1.0" encoding="UTF-8"?>\n<root>\n')
+
+        with open(csv_file, mode='r', newline='', encoding='utf-8') as csvfile:
+            reader = csv.DictReader(csvfile, delimiter=',')
+            headers = reader.fieldnames
+
+            chunk = []
+            for line in reader:
+                chunk.append(line)
+
+                # saves chunk if it's full
+                if len(chunk) >= chunk_size:
+                    
+                    for item in chunk:
+                        xmlfile.write('  <' + objname + '>\n')
+                        for field in headers:
+                            # process special chars
+                            value = str(item[field]) if item[field] is not None else ''
+                            escaped_value = saxutils.escape(value)
+                            xmlfile.write(f'    <{field}>{escaped_value}</{field}>\n')
+                        xmlfile.write('  </' + objname + '>\n')
+                    
+                    # reset chunk
+                    chunk = []
+
+            # leftover chunk (less than 10k lines)
+            if chunk:
+                for item in chunk:
+                    xmlfile.write('  <' + objname + '>\n')
+                    for field in headers:
+                        value = str(item[field]) if item[field] is not None else ''
+                        escaped_value = saxutils.escape(value)
+                        xmlfile.write(f'    <{field}>{escaped_value}</{field}>\n')
+                    xmlfile.write('  </' + objname + '>\n')
+
+        xmlfile.write('</root>')
 
 def fill_empty_fields(xml_file):
     tree = ET.parse(xml_file)
     root = tree.getroot()
 
-    for airport in root.findall("Airport"):
+    for airport in root.findall("Temp"):
         for field in airport:
             if field.text is None or field.text.strip() == "":
-                field.text = "0"
+                field.text = "null"
 
     tree.write(xml_file, encoding='utf-8', xml_declaration=True)
 
@@ -59,9 +86,7 @@ def validate_xml(xml_file, xsd_file):
         return False
 
 
-
-
-
+# ============== gRPC Services ==============
 
 class SendFileService(server_services_pb2_grpc.SendFileServiceServicer):
 
@@ -153,6 +178,86 @@ class ImporterService(server_services_pb2_grpc.ImporterServiceServicer):
     def __init__(self, *args, **kwargs):
         pass
 
+    def UploadCSV(self, request, context):
+        os.makedirs(MEDIA_PATH, exist_ok=True)
+        file_path = os.path.join(MEDIA_PATH, "./csv/", request.file_name + request.file_mime)
+
+        ficheiro_em_bytes = request.file
+
+        with open(file_path, 'wb') as f:
+            f.write(ficheiro_em_bytes)
+        
+        logger.info(f"{DBHOST}:{DBPORT}", exc_info=True)
+        # Establish connection to PostgreSQL
+        try:
+            # Connect to the database
+            conn = pg8000.connect(user=f'{DBUSERNAME}', password=f'{DBPASSWORD}', host=f'{DBHOST}', port=f'{DBPORT}', database=f'{DBNAME}')
+            cursor = conn.cursor()
+            
+            #SQL query to create a table
+            create_table_query = """
+            CREATE TABLE IF NOT EXISTS data (
+                id SERIAL PRIMARY KEY,
+                ident VARCHAR(100),
+                type VARCHAR(100) NOT NULL,
+                name VARCHAR(100) NOT NULL,
+                latitude_deg FLOAT,
+                longitude_deg FLOAT,
+                elevation_ft INT,
+                continent VARCHAR(100),
+                iso_country VARCHAR(100),
+                iso_region VARCHAR(100),
+                municipality VARCHAR(100),
+                scheduled_service VARCHAR(100),
+                gps_code VARCHAR(100),
+                local_code VARCHAR(100)
+            );
+            """
+
+            # Read the CSV file
+            with open(file_path, 'r') as f:
+                reader = csv.DictReader(f)
+                next(reader)  # Skip the header row
+                for row in reader:
+
+                    # SQL query to insert data into the table
+                    insert_query = """
+                    INSERT INTO data (ident, type, name, latitude_deg, longitude_deg, elevation_ft, 
+                    continent, iso_country, iso_region, municipality, scheduled_service, gps_code, local_code)
+                    SELECT %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s
+                    WHERE NOT EXISTS (
+                        SELECT 1 FROM data WHERE ident = %s
+                    );
+                    """
+
+                    cursor.execute(insert_query, (
+                        row['ident'],
+                        row['type'],
+                        row['name'],
+                        float(row['latitude_deg']) if row['latitude_deg'] else None,
+                        float(row['longitude_deg']) if row['longitude_deg'] else None,
+                        int(row['elevation_ft']) if row['elevation_ft'] else None,
+                        row['continent'],
+                        row['iso_country'],
+                        row['iso_region'],
+                        row['municipality'],
+                        row['scheduled_service'],
+                        row['gps_code'] if row['gps_code'] else None,
+                        row['local_code'] if row['local_code'] else None,
+                        row['ident']
+                    ))
+
+            # Execute the SQL query to create the table
+            cursor.execute(create_table_query)
+            # Commit the changes (optional in this case since it's a DDL query)
+            conn.commit()
+            return server_services_pb2.FileUploadResponse(success=True)
+        except Exception as e:
+            logger.error(f"Error: {str(e)}", exc_info=True)
+            context.set_details(f"Failed: {str(e)}")
+            context.set_code(grpc.StatusCode.INTERNAL)
+            return server_services_pb2.FileUploadResponse(success=False)   
+
     def ConvertToXML(self, request, context):
         try:
            # Ensure absolute paths are used
@@ -180,10 +285,10 @@ class ImporterService(server_services_pb2_grpc.ImporterServiceServicer):
                 logger.error(f"Input file not found: {input_csv_path}")
                 context.set_details("Input file not found")
                 context.set_code(grpc.StatusCode.NOT_FOUND)
-                return server_services_pb2.XMLFileResponse()
+                return server_services_pb2.XMLFileResponse(success=False)
 
             # Convert CSV to XML
-            csv_to_xml(input_csv_path, xml_output_path, "Airport")
+            csv_to_xml(input_csv_path, xml_output_path, "Temp")
             
             # Fill empty fields
             fill_empty_fields(xml_output_path)
@@ -193,26 +298,25 @@ class ImporterService(server_services_pb2_grpc.ImporterServiceServicer):
             if os.path.exists(xsd_path):
                 if not validate_xml(xml_output_path, xsd_path):
                     logger.warning(f"XML validation failed for {xml_output_path}")
-                else: 
-                    logger.info(f"XML validation passed for {xml_output_path}")
+                            
+            logger.info(f"XML validation passed for {xml_output_path}")
 
             
-            # Read the XML file content
-            with open(xml_output_path, 'rb') as xml_file:
-                xml_content = xml_file.read()
-            
-            return server_services_pb2.XMLFileResponse(xml_file=xml_content)
+            return server_services_pb2.XMLFileResponse(success=True)
         except Exception as e:
             logger.error(f"XML Conversion Error: {str(e)}", exc_info=True)
-            context.set_details(f"Conversion failed: {str(e)}")
+            context.set_details(f"Failed: {str(e)}")
             context.set_code(grpc.StatusCode.INTERNAL)
-            return server_services_pb2.XMLFileResponse()
+            return server_services_pb2.XMLFileResponse(success=False)
         
     def CheckConversionStatus(self, request, context):
         pass
 
     def DownloadXML(self, request, context):
         pass
+
+
+# ============= main ==============
 
 def serve():
 
@@ -229,7 +333,6 @@ def serve():
 
     print(f'Server running at {GRPC_SERVER_PORT}...')
     server.wait_for_termination()
-
 
 if __name__ == '__main__':
     serve()
